Refactor for Sonar smells and code coverage
Various code formatting and style fixes (to address Sonar code smells).

Introduce a BabelServiceClientException to replace the use of the base
Exception class.

Make both the BabelServiceClient class and its Factory class use an
interface, to allow for Spring injection (using annotations) and for
simplified mocking.

Refactor the Notification Publishing by making this the responsibility
of the main Service classes and not the deployment manager.

Issue-ID: AAI-1207
Change-Id: I6bf8a504fa7b98e9158c9976fd1d88c6c54027c1
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>