Update patch set 1
Patch Set 1:

(1 comment)

This new JUnit test calls only one of the five existing getter methods in BasicStatusMsg.java and therefore increases the code coverage of this class from 18.8% to 31.2%
If we are creating a new test class simply to increase the code coverage, why not add another few lines of code in the same commit in order to reach 100%?
Even better would be to NOT construct this class (which exists purely for implementation purposes) and to instead create a functional test that exercises one or more of the Builder APIs (see CompDoneStatusMessageBuilder.java). This decouples the unit tests from the implementation detail, and reduces some of the maintenance overhead for future developers.

Patch-set: 1