Update patch set 1
Patch Set 1:

Hi Paul.  I think this is the first time we've used Alpine as the base image in one of the AAI services.  What type of testing/validation did you do?

I'm 100% on-board with reducing the image size.  As long as you've done some validation to ensure the service starts up and attempts to connect with SDC, I'm happy to give a +2.

Patch-set: 1