Update patch set 1
Patch Set 1:

Hi Steven. The CSIT jobs are not very easy to run locally so we are basing on each team to guide with the tests as we can't know the logic for all project teams, so if you can give hints on how to run this image and test this I will be happy to put it into practice.
If I simply run the container it complains that:
CONFIG_HOME must be set in order to start up the process

Patch-set: 1