Update patch set 1
Patch Set 1:

Hi Steven, Can you please expand on this nss dependency? is this a package we can preinstall ? or is something you install at runtime in your environemnt ? I am just trying to undestand if this is a solvable issue or not. For images that are not portable to alpine we abandon the task but we need to undestand if there is a strong reason for not migrating them.

Patch-set: 1
Reviewer: Gerrit User 3621 <3621@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0