Update patch set 1
Patch Set 1:

> Hi Steven, Can you please expand on this nss dependency? is this a
 > package we can preinstall ? or is something you install at runtime
 > in your environemnt ? I am just trying to undestand if this is a
 > solvable issue or not. For images that are not portable to alpine
 > we abandon the task but we need to undestand if there is a strong
 > reason for not migrating them.

I don't see any reason why we couldn't pre-install this.  To be honest, I haven't looked at what uses this package.  I imagine that there are a bunch of packages which come with the ubuntu image that aren't by default installed on Alpine.  Will probably be trial and error to determine which ones we'll need to pre-install on alpine.  Which means you'll need to find a way to test the images before you push to gerrit.  I typically use OOM / kubernetes as my testing environment.  There should be instructions on the ONAP wiki under the OOM project for doing this.

Patch-set: 1